break down components file into different components
