+++ apps/nameguard.io/src/app/components/atoms/icons/Error404IconLarge.tsx
@@ -1,0 +1,52 @@
+export const Error404IconLarge = () => (
+  <svg
+    width="500"
+    height="200"
+    viewBox="0 0 500 200"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    className="hidden md:block flex-shrink-0"
+  >
+    <g clipPath="url(#clip0_1447_5217)">
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M-0.468079 132.061V159.582H78.8675V188.637H112.281V159.582H132.78V131.577H112.281V23.3477H92.265H79.5132H68.537L-0.468079 132.061ZM79.5132 61.4417H78.2219L34.7205 130.285V131.577H79.5132V61.4417ZM367 132.061V159.582H446.336V188.637H479.749V159.582H500.248V131.577H479.749V23.3477H459.733H446.981H436.005L367 132.061ZM446.981 61.4417H445.69L402.189 130.285V131.577H446.981V61.4417Z"
+        fill="url(#paint0_linear_1447_5217)"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M255 5.37109L265.297 15.0931C280.541 29.4844 301.093 38.2966 323.75 38.2966C324.111 38.2966 324.472 38.2944 324.832 38.2899L335.821 38.1539L339.263 48.5906C342.99 59.891 345 71.9513 345 84.4503C345 137.945 308.302 182.809 258.728 195.529L255 196.486L251.272 195.529C201.698 182.809 165 137.945 165 84.4503C165 71.9513 167.01 59.891 170.737 48.5906L174.179 38.1539L185.168 38.2899C185.528 38.2944 185.889 38.2966 186.25 38.2966C208.907 38.2966 229.459 29.4844 244.703 15.0931L255 5.37109ZM196.632 67.8349C195.562 73.1992 195 78.7537 195 84.4503C195 122.512 220.224 154.78 255 165.423C289.776 154.78 315 122.512 315 84.4503C315 78.7537 314.438 73.1992 313.368 67.8349C291.604 65.8903 271.572 57.8932 255 45.5521C238.428 57.8932 218.396 65.8903 196.632 67.8349Z"
+        fill="url(#paint1_linear_1447_5217)"
+      />
+    </g>
+    <defs>
+      <linearGradient
+        id="paint0_linear_1447_5217"
+        x1="249.89"
+        y1="23.3477"
+        x2="249.89"
+        y2="188.637"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stopColor="#4F4F4F" />
+        <stop offset="1" />
+      </linearGradient>
+      <linearGradient
+        id="paint1_linear_1447_5217"
+        x1="255"
+        y1="5.37109"
+        x2="255"
+        y2="196.486"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stopColor="#4F4F4F" />
+        <stop offset="1" />
+      </linearGradient>
+      <clipPath id="clip0_1447_5217">
+        <rect width="500" height="200" fill="white" />
+      </clipPath>
+    </defs>
+  </svg>
+);
