+++ apps/nameguard.io/src/app/components/atoms/icons/Error404IconSmall.tsx
@@ -1,0 +1,35 @@
+export const Error404IconSmall = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="221"
+    height="88"
+    viewBox="0 0 221 88"
+    fill="none"
+    className="block md:hidden flex-shrink-0"
+  >
+    <g clipPath="url(#clip0_1450_6699)">
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M117.231 6.64097L112.7 2.36328L108.169 6.64097C101.462 12.9731 92.4191 16.8505 82.45 16.8505C82.2911 16.8505 82.1324 16.8495 81.974 16.8476L77.1389 16.7877L75.6244 21.3799C73.9845 26.352 73.1 31.6586 73.1 37.1581C73.1 60.6956 89.247 80.4361 111.06 86.033L112.7 86.4538L114.34 86.033C136.153 80.4361 152.3 60.6956 152.3 37.1581C152.3 31.6586 151.415 26.352 149.776 21.3799L148.261 16.7877L143.426 16.8476C143.268 16.8495 143.109 16.8505 142.95 16.8505C132.981 16.8505 123.938 12.9731 117.231 6.64097ZM86.3 37.1581C86.3 34.6516 86.5475 32.2077 87.0182 29.8473C96.5942 28.9917 105.408 25.473 112.7 20.0429C119.992 25.473 128.806 28.9917 138.382 29.8473C138.853 32.2077 139.1 34.6516 139.1 37.1581C139.1 53.9054 128.002 68.1033 112.7 72.786C97.3984 68.1033 86.3 53.9054 86.3 37.1581ZM0.294037 70.2163V58.1069L30.6563 10.2731H35.4858H41.0966H49.9034V57.8938H58.9233V70.2163H49.9034V83.0004H35.2017V70.2163H0.294037ZM34.9176 27.0344H35.4858V57.8938H15.777V57.3256L34.9176 27.0344ZM161.98 70.2163V58.1069L192.342 10.2731H197.172H202.783H211.589V57.8938H220.609V70.2163H211.589V83.0004H196.888V70.2163H161.98ZM196.604 27.0344H197.172V57.8938H177.463V57.3256L196.604 27.0344Z"
+        fill="url(#paint0_linear_1450_6699)"
+      />
+    </g>
+    <defs>
+      <linearGradient
+        id="paint0_linear_1450_6699"
+        x1="110.452"
+        y1="2.36328"
+        x2="110.452"
+        y2="86.4538"
+        gradientUnits="userSpaceOnUse"
+      >
+        <stop stopColor="#4F4F4F" />
+        <stop offset="1" />
+      </linearGradient>
+      <clipPath id="clip0_1450_6699">
+        <rect width="220" height="88" fill="white" transform="translate(0.5)" />
+      </clipPath>
+    </defs>
+  </svg>
+);
